print all

+++ 0x10-variadic_functions/3-print_all.c
@@ -27,16 +27,20 @@ void print_all(const char * const format, ...)
 		switch (format[j])
 		{
 			case 'c':
-				printf("%c", va_arg(args, int)); c = 1;
+				printf("%c", va_arg(args, int)); 
+				c = 1;
 				break;
 			case 'i':
-				printf("%d", va_arg(args, int)); c = 1;
+				printf("%d", va_arg(args, int)); 
+				c = 1;
 				break;
 			case 'f':
-				printf("%f", va_arg(args, int)); c = 1;
+				printf("%f", va_arg(args, double)); 
+				c = 1;
 				break;
 			case 's':
-				str = va_arg(args, char*); c = 1;
+				str = va_arg(args, char*); 
+				c = 1;
 				if (str == NULL)
 					printf("(nil)");
 				printf("%s", str);
